Z-Image-sever/requirements.txt
# Z-Image Server 依赖
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
httpx>=0.26.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# 数据库
sqlalchemy>=2.0.25
aiosqlite>=0.19.0

# 认证
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# 工具
python-dotenv>=1.0.0
aiofiles>=23.2.1
Pillow>=10.0.0



